+++ addon-test-support/-private/converters.js
@@ -5,5 +5,3 @@ export function element(label, next) {
 
   next(null, collectionOrElement);
 }
-
-