+++ addon-test-support/-private/then-steps.js
@@ -23,12 +23,12 @@ const steps = {
     expect(currentURL()).to.equal(url);
   },
 
-  "Then there should be (?:(\\d+) )?$element"(countRaw = "1", [collection]) {
+  "Then there should be (?:(\\d+) )?$opinionatedElement"(countRaw = "1", [collection]) {
     const count = parseInt(countRaw, 10);
     expect(collection, this.step).to.have.length(count);
   },
 
-  "Then $element should (?:have text|say) \"$text\""([collection, label, selector], text) {
+  "Then $opinionatedElement should (?:have text|say) \"$text\""([collection, label, selector], text) {
     assert(`Expected a single element, but ${collection.length} found.\nLabel: ${label}\nSelector: ${selector}\nStep: ${this.step}`, collection.length === 1);
     expect(collection[0].textContent.trim(), this.step).equal(text);
   },