+++ addon-test-support/index.js
@@ -1,4 +1,4 @@
-import composeSteps from 'ember-cli-yadda-opinionated/test-support/-private/compose';
+import composeSteps from 'ember-cli-yadda-opinionated/test-support/-private/compose-steps';
 import { element } from 'ember-cli-yadda-opinionated/test-support/-private/converters';
 
 import {
@@ -8,6 +8,7 @@ import {
 import { labelMap, transformsMap } from 'ember-cli-yadda-opinionated/test-support/-private/maps';
 import findBylabel from 'ember-cli-yadda-opinionated/test-support/-private/find-by-label';
 import selectorFromLabel from 'ember-cli-yadda-opinionated/test-support/-private/selector-from-label';
+import setupDictionary from 'ember-cli-yadda-opinionated/test-support/-private/setup-dictionary';
 import givenSteps from 'ember-cli-yadda-opinionated/test-support/-private/given-steps';
 import thenSteps from 'ember-cli-yadda-opinionated/test-support/-private/then-steps';
 import whenSteps from 'ember-cli-yadda-opinionated/test-support/-private/when-steps';
@@ -20,6 +21,7 @@ export {
   labelMap,
   REGEX_LABEL,
   selectorFromLabel,
+  setupDictionary,
   thenSteps,
   transformsMap,
   whenSteps